fix(snapshot): improve snapshot index validation
Refactored restoreSnapshot to use size_t for the index parameter and
updated the index range check to use SNAPSHOT_COUNT instead of a magic
number. Added an additional check to ensure the snapshot exists before
restoring, throwing a runtime error if it does not. This improves
robustness and prevents potential out-of-bounds or null dereference
errors.

+++ lib/SnapshotManager.hpp
@@ -12,5 +12,5 @@ private:
     size_t count = 0;
 public:
     void createSnapshot(const Dispatcher& globalDispatch);
-    void restoreSnapshot(int index, Dispatcher& globalDispatch) const;
+    void restoreSnapshot(size_t index, Dispatcher& globalDispatch) const;
 };
